(BSR)[PRO] refactor: Dialog with new archi
