+++ pro/src/components/ArchiveConfirmationModal/ArchiveConfirmationModal.tsx
@@ -7,7 +7,7 @@ import {
 } from 'apiClient/v1'
 import { useAnalytics } from 'app/App/analytics/firebase'
 import { Events } from 'commons/core/FirebaseEvents/constants'
-import { ConfirmDialog } from 'components/Dialog/ConfirmDialog/ConfirmDialog'
+import { ConfirmDialog } from 'components/ConfirmDialog/ConfirmDialog'
 import strokeThingIcon from 'icons/stroke-thing.svg'
 
 interface OfferEducationalModalProps {
