+++ pro/src/components/CollectiveOffersTable/CollectiveOfferRow/CollectiveActionsCells/DuplicateOfferDialog/DuplicateOfferDialog.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 
-import { ConfirmDialog } from 'components/Dialog/ConfirmDialog/ConfirmDialog'
+import { ConfirmDialog } from 'components/ConfirmDialog/ConfirmDialog'
 import { BaseCheckbox } from 'ui-kit/form/shared/BaseCheckbox/BaseCheckbox'
 
 import styles from './DuplicateOfferDialog.module.scss'
