+++ pro/src/components/ConfirmDialog/ConfirmDialog.spec.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
 
-import { ConfirmDialog } from '../ConfirmDialog'
+import { ConfirmDialog } from './ConfirmDialog'
 
 describe('ConfirmDialog', () => {
   const onConfirm = vi.fn()
