+++ pro/src/components/Dialog/Dialog.module.scss
