+++ pro/src/components/IndividualOffer/DialogStockDeleteConfirm/DialogStockEventDeleteConfirm.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { ConfirmDialog } from 'components/Dialog/ConfirmDialog/ConfirmDialog'
+import { ConfirmDialog } from 'components/ConfirmDialog/ConfirmDialog'
 import strokeTrashIcon from 'icons/stroke-trash.svg'
 
 import styles from './DialogStockDeleteConfirm.module.scss'
