+++ pro/src/components/IndividualOffer/DialogStockDeleteConfirm/DialogStockThingDeleteConfirm.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { ConfirmDialog } from 'components/Dialog/ConfirmDialog/ConfirmDialog'
+import { ConfirmDialog } from 'components/ConfirmDialog/ConfirmDialog'
 import strokeTrashIcon from 'icons/stroke-trash.svg'
 
 interface DialogStockDeleteConfirmProps {
