+++ pro/src/components/IndividualOffer/PriceCategoriesScreen/PriceCategoriesScreen.tsx
@@ -12,7 +12,7 @@ import { isOfferDisabled } from 'commons/core/Offers/utils/isOfferDisabled'
 import { isOfferAllocineSynchronized } from 'commons/core/Offers/utils/typology'
 import { useNotification } from 'commons/hooks/useNotification'
 import { useOfferWizardMode } from 'commons/hooks/useOfferWizardMode'
-import { ConfirmDialog } from 'components/Dialog/ConfirmDialog/ConfirmDialog'
+import { ConfirmDialog } from 'components/ConfirmDialog/ConfirmDialog'
 import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferNavigation/constants'
 import { RouteLeavingGuardIndividualOffer } from 'components/RouteLeavingGuardIndividualOffer/RouteLeavingGuardIndividualOffer'
 import { ActionBar } from 'pages/IndividualOffer/components/ActionBar/ActionBar'
