+++ pro/src/components/IndividualOffer/StocksThing/ActivationCodeFormDialog/ActivationCodeFormDialog.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useState } from 'react'
 
-import { Dialog } from 'components/Dialog/Dialog/Dialog'
+import { Dialog } from 'components/Dialog/Dialog'
 import strokeCodeIcon from 'icons/stroke-code.svg'
 
 import {
