+++ pro/src/components/RedirectDialog/RedirectDialog.module.scss
