+++ pro/src/components/RedirectToBankAccountDialog/RedirectToBankAccountDialog.tsx
@@ -2,8 +2,8 @@ import { useNavigate } from 'react-router-dom'
 
 import { useAnalytics } from 'app/App/analytics/firebase'
 import { Events, VenueEvents } from 'commons/core/FirebaseEvents/constants'
-import { RedirectDialog } from 'components/Dialog/RedirectDialog/RedirectDialog'
 import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferNavigation/constants'
+import { RedirectDialog } from 'components/RedirectDialog/RedirectDialog'
 import fullWaitIcon from 'icons/full-wait.svg'
 import strokePartyIcon from 'icons/stroke-party.svg'
