+++ pro/src/pages/Homepage/components/Offerers/Offerers.tsx
@@ -6,13 +6,12 @@ import { useAnalytics } from 'app/App/analytics/firebase'
 import { Events } from 'commons/core/FirebaseEvents/constants'
 import { SelectOption } from 'commons/custom_types/form'
 import { useActiveFeature } from 'commons/hooks/useActiveFeature'
-import { RedirectDialog } from 'components/Dialog/RedirectDialog/RedirectDialog'
+import { Card } from 'components/Card/Card'
+import { RedirectDialog } from 'components/RedirectDialog/RedirectDialog'
 import { SoftDeletedOffererWarning } from 'components/SoftDeletedOffererWarning/SoftDeletedOffererWarning'
 import fullWaitIcon from 'icons/full-wait.svg'
 import strokePartyIcon from 'icons/stroke-party.svg'
 import { Spinner } from 'ui-kit/Spinner/Spinner'
-
-import { Card } from '../../../../components/Card/Card'
 
 import { OffererCreationLinks } from './components/OffererCreationLinks/OffererCreationLinks'
 import { PartnerPages } from './components/PartnerPages/PartnerPages'
