+++ pro/src/pages/Reimbursements/BankInformations/AddBankInformationsDialog.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { useAnalytics } from 'app/App/analytics/firebase'
 import { BankAccountEvents } from 'commons/core/FirebaseEvents/constants'
 import { DS_BANK_ACCOUNT_PROCEDURE_ID } from 'commons/utils/config'
-import { Dialog } from 'components/Dialog/Dialog/Dialog'
+import { Dialog } from 'components/Dialog/Dialog'
 import fullLinkIcon from 'icons/full-link.svg'
 import strokeLinkIcon from 'icons/stroke-link.svg'
 import { ButtonLink } from 'ui-kit/Button/ButtonLink'
