+++ pro/src/pages/Signup/SignupContainer/MaybeAppUserDialog/MaybeAppUserDialog.tsx
@@ -1,7 +1,6 @@
-import './MaybeAppUserDialog.scss'
-
-import { RedirectDialog } from 'components/Dialog/RedirectDialog/RedirectDialog'
+import { RedirectDialog } from 'components/RedirectDialog/RedirectDialog'
 import strokeFraudIcon from 'icons/stroke-fraud.svg'
+import './MaybeAppUserDialog.scss'
 
 export const MaybeAppUserDialog = ({
   onCancel,
