+++ pro/src/pages/VenueSettings/VenueProvidersManager/StocksProviderForm/StocksProviderForm.tsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react'
 import { PostVenueProviderBody } from 'apiClient/v1'
 import { useAnalytics } from 'app/App/analytics/firebase'
 import { SynchronizationEvents } from 'commons/core/FirebaseEvents/constants'
-import { ConfirmDialog } from 'components/Dialog/ConfirmDialog/ConfirmDialog'
+import { ConfirmDialog } from 'components/ConfirmDialog/ConfirmDialog'
 import fullLinkIcon from 'icons/full-link.svg'
 import strokeConnectIcon from 'icons/stroke-connect.svg'
 import { Button } from 'ui-kit/Button/Button'
